+++ banzai/qc/header_checker.py
@@ -6,6 +6,7 @@ format and validates their values.
 import logging
 
 from banzai.stages import Stage
+from banzai.utils import qc
 
 logger = logging.getLogger(__name__)
 
@@ -92,7 +93,7 @@ class HeaderSanity(Stage):
             qc_results["header.keywords.missing.names"] = missing_keywords
         if are_keywords_na:
             qc_results["header.keywords.na.names"] = na_keywords
-        self.save_qc_results(qc_results, image)
+        qc.save_qc_results(self.pipeline_context, qc_results, image)
         return missing_keywords + na_keywords
 
     def check_ra_range(self, image, bad_keywords=None):
@@ -116,8 +117,9 @@ class HeaderSanity(Stage):
             if is_bad_ra_value:
                 sentence = 'The header CRVAL1 key got the unexpected value : {0}'.format(ra_value)
                 logger.error(sentence, image=image)
-            self.save_qc_results({"header.ra.failed": is_bad_ra_value,
-                                  "header.ra.value": ra_value}, image)
+            qc_results = {"header.ra.failed": is_bad_ra_value,
+                          "header.ra.value": ra_value}
+            qc.save_qc_results(self.pipeline_context, qc_results, image)
 
     def check_dec_range(self, image, bad_keywords=None):
         """
@@ -140,8 +142,9 @@ class HeaderSanity(Stage):
             if is_bad_dec_value:
                 sentence = 'The header CRVAL2 key got the unexpected value : {0}'.format(dec_value)
                 logger.error(sentence, image=image)
-            self.save_qc_results({"header.dec.failed": is_bad_dec_value,
-                                  "header.dec.value": dec_value}, image)
+            qc_results = {"header.dec.failed": is_bad_dec_value,
+                          "header.dec.value": dec_value}
+            qc.save_qc_results(self.pipeline_context, qc_results, image)
 
     def check_exptime_value(self, image, bad_keywords=None):
         """
@@ -166,4 +169,4 @@ class HeaderSanity(Stage):
                                'null or negative value'.format(exptime_value)
                     logger.error(sentence, image=image)
                 qc_results["header.exptime.failed"] = is_exptime_null
-            self.save_qc_results(qc_results, image)
+            qc.save_qc_results(self.pipeline_context, qc_results, image)