+++ banzai/tests/test_dark_comparer.py
@@ -46,8 +46,7 @@ def test_raises_an_exception_if_ny_are_different(mock_cal):
 
 
 @mock.patch('banzai.calibrations.ApplyCalibration.get_calibration_filename')
-@mock.patch('banzai.stages.Stage.save_qc_results')
-def test_flags_bad_if_no_master_calibration(mock_save_qc, mock_cal):
+def test_flags_bad_if_no_master_calibration(mock_cal):
     mock_cal.return_value = None
     context = FakeContext()
     context.FRAME_CLASS = FakeDarkImage
@@ -57,8 +56,7 @@ def test_flags_bad_if_no_master_calibration(mock_save_qc, mock_cal):
 
 
 @mock.patch('banzai.calibrations.ApplyCalibration.get_calibration_filename')
-@mock.patch('banzai.stages.Stage.save_qc_results')
-def test_does_not_flag_noisy_images(mock_save_qc, mock_cal, set_random_seed):
+def test_does_not_flag_noisy_images(mock_cal, set_random_seed):
     mock_cal.return_value = 'test.fits'
     master_dark_fraction = 0.05
     nx = 101
@@ -81,8 +79,7 @@ def test_does_not_flag_noisy_images(mock_save_qc, mock_cal, set_random_seed):
 
 
 @mock.patch('banzai.calibrations.ApplyCalibration.get_calibration_filename')
-@mock.patch('banzai.stages.Stage.save_qc_results')
-def test_does_flag_bad_images(mock_save_qc, mock_cal, set_random_seed):
+def test_does_flag_bad_images( mock_cal, set_random_seed):
     mock_cal.return_value = 'test.fits'
     master_dark_fraction = 0.05
     nx = 101