+++ CHANGES.md
@@ -1,3 +1,8 @@
+0.18.4 (2019-02-07)
+-------------------
+- Moved the function that adds quality control information to ElasticSearch
+  outside of Stage class and to the quality control utils
+
 0.18.3 (2019-02-05)
 -------------------
 - Fixed query for master calibration to work even when block start